Make two versions of the error management code
The first one, activated by default by the "verbose-errors" feature,
can make an error chain by accumulating errors and input position
while backtracing in a parse tree, and show precise information on
which parts of the input were matched by which node in the parse
tree.

The second one will only return a simple error code. It is less
flexible, but will not keep references to the input, and it will
not put previous errors in a Box, thus reducing the code size
(less error drops) et and making the parsers faster.

On some parsers, the simpler solution got a 30% gain in performance,
returning to the speed nom had pre 1.0