do_parse is the replacement for chain!
it keeps the sequential application of parsers, the ability to
bind the result of a parser to a name, and reuse results from
previous parsers in following ones.

It removes the mutable bindings (if a value must be modified,
bind it to a new one and use value!()), and the '?' for optional
parsers, which can be done with opt!().

To create a struct from the assembled result, use map!() to get
the resulting tuple and pass it to a closure or a 'new' function