Only mark FileProducer as Eof on full buffer
Previously, if `FileProducer.refill()` was called twice in a row the second
would read 0 additional bytes and assume that was due to Eof when really there
was just no more room left in the buffer.

When `FileProducer.apply()` returns `ConsumerState::Continue(Move::Await(_))`
it could be due to either 1) the buffer needing a refill or 2) the buffer not
being large enough.  There's no way for the caller to differentiate between
these states.

A sensisible plan may be to try to `apply()` again, in the case of (1) it will
`refill()` the buffer and continue on it's way.  However, if it's (2) then the
buffer will already be full `refill()` will read 0 bytes and erroneously mark
the producer at Eof.  There is no way to reset or recover from this condition.

Now that `refill()` has been fixed, the caller can respond to subsequent
`Move::Await(_)`s by `resize()`ing the producer's buffer and trying again.