store the error chain in a vec instead of a Box based linked list
having each node in the error chain contain a pointer to the next
error resulted in a lot of allocations. Now we allocate a Vec to
contain the errors, and append them when necessary.

This gives a nearly 10% performance boost over the previous approach.
It is still slower than the simple-errors feature