Removed leading Zeros in Ranges
Removing leading zeroes in ranges helped making the code a bit
more intuitive. Also this made Range<usize> as a trait
superflous in some cases.

ATTENTION: This also ment in two case to switch from RangeFrom to
RangeFull. This might break some code which does not rely on `str`
ot `[T]` for which the Slice trait is implemented for all kinds of
Ranges by nom.