Revert "Merge branch 'master' into new-many"
This reverts commit 4ed25b1ed693f7d27a7eba14f97e881fe740dc05, reversing
changes made to 6b7c938740f53bee0458e89812fdcd5f4a913e4f.

This was a bad merge, since some new feature `ParseTo` was coming in
which is not compatible with the modifications I made for my new macros.
Will need someone aware of how the new feature works to merge this manually.