Add inline attribute to the to_usize functions
The nom take parser automatically converts it's argument to usize
using to_usize.

However, if the argument is already usize, this should be a no-op and
the compiler should just skip the call. Surprisingly, during my tests,
I noticed that a large part of the cycles were wasted in
to_usize(usize).

Adding the [inline] attribute fixes it for me.

Example:

Before the change, the parser would take 29,606,796,156 cycles.
After the change, the parser only takes 9,245,008,172 cycles.
This translates in runtime decreasing from 7.6 seconds to 2.8 seconds.