doc: wording in choosing_a_combinator
I took a stab at introducing some new wording.
- Surfaces `and_then` and its functional variant `map_parser` which I didn't find while first learning Nom.
- Attempts to provide some clarity that I found was lacking around `flat_map`
- Reorders the Modifiers list to try and draw parallels between Trait and Combinator versions.

While I was getting up to speed on Nom, discovering the correct combinator was much harder than discovering a trait function. This was difficult as the example code and explanation usually lives on the combinator. This lead to me using the couple of combinators and base parsers introduced through tutorials and examples, and getting stuck a lot reading this page and things not fitting in the slots like I expected.

Happy for feedback and to change things to fit. If this feels like the right kind of wording, I may go in and update trait function documentation to make things a little friendlier for discovery of the related combinators through intellisense. Also feel free to tell me to go make a blog post instead.